feat: 예산 추천 기능 추가

+++ src/main/java/com/wanted/budgetmanagement/api/budget/dto/BudgetSettingRequest.java
@@ -19,7 +19,7 @@ public class BudgetSettingRequest {
 
     @Schema(description = "설정 예산", example = "100000")
     @NotNull(message = "예산을 입력해주세요.")
-    private int money;
+    private long money;
 
     @Schema(description = "예산 카테고리", example = "식비")
     @NotBlank(message = "카테고리를 입력해주세요")
